@@ -1,4 +1,4 @@
-import React, { FC } from 'react'
+import { FC } from 'react'
 
 import Footer from '@/components/organisms/Footer'
 import Navbar from '@/components/organisms/Navbar'
