@@ -3,7 +3,7 @@ import clsx from 'clsx'
 import Head from 'next/head'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-import React, { FC } from 'react'
+import { FC } from 'react'
 
 import Container from '@/components/atoms/Container'
 import Hero, { HeroTitle } from '@/components/organisms/Hero'
