@@ -3,7 +3,7 @@ import chunk from 'lodash.chunk'
 import { GetServerSideProps, NextPage } from 'next'
 import Link from 'next/link'
 import { NextSeo } from 'next-seo'
-import React, { useCallback } from 'react'
+import { useCallback } from 'react'
 import InfiniteScroll from 'react-infinite-scroll-component'
 import { useSWRInfinite } from 'swr'
 import { shareCard } from '@/assets'
